app/requirements.txt
# =============================================================================
# NTO - DEPENDÊNCIAS PYTHON
# =============================================================================
# Data: 22/12/2025
# Versão: 2.1 - Arquitetura autossuficiente

# -----------------------------------------------------------------------------
# CORE DJANGO
# -----------------------------------------------------------------------------
Django==4.2.8
django-environ==0.11.2
python-decouple==3.8

# -----------------------------------------------------------------------------
# BANCO DE DADOS
# -----------------------------------------------------------------------------
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# -----------------------------------------------------------------------------
# CACHE E FILAS
# -----------------------------------------------------------------------------
redis==5.0.1
celery==5.3.4
django-redis==5.4.0

# -----------------------------------------------------------------------------
# APIS E SERIALIZAÇÃO
# -----------------------------------------------------------------------------
djangorestframework==3.14.0
django-cors-headers==4.3.1

# -----------------------------------------------------------------------------
# AUTENTICAÇÃO E SEGURANÇA
# -----------------------------------------------------------------------------
django-allauth==0.57.0
django-crispy-forms==2.1
crispy-tailwind==0.5.0

# -----------------------------------------------------------------------------
# ARQUIVOS E STORAGE
# -----------------------------------------------------------------------------
django-storages==1.14.2
boto3==1.34.0
Pillow==10.1.0

# -----------------------------------------------------------------------------
# FRONTEND
# -----------------------------------------------------------------------------
django-compressor==4.4
django-widget-tweaks==1.5.0

# -----------------------------------------------------------------------------
# DESENVOLVIMENTO
# -----------------------------------------------------------------------------
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# -----------------------------------------------------------------------------
# MONITORAMENTO
# -----------------------------------------------------------------------------
sentry-sdk[django]==1.38.0

# -----------------------------------------------------------------------------
# INTEGRAÇÕES IA
# -----------------------------------------------------------------------------
openai==1.6.1
google-generativeai==0.3.2
httpx==0.25.2

# -----------------------------------------------------------------------------
# WEB SCRAPING E TRENDS
# -----------------------------------------------------------------------------
beautifulsoup4==4.12.2
lxml==4.9.3
pytrends==4.9.2
playwright==1.40.0

# -----------------------------------------------------------------------------
# UTILITÁRIOS
# -----------------------------------------------------------------------------
python-dateutil==2.8.2
requests==2.31.0
validators==0.22.0
imagehash==4.3.1

# -----------------------------------------------------------------------------
# SERVIDOR
# -----------------------------------------------------------------------------
gunicorn==21.2.0
whitenoise==6.6.0
